Added initial page structure and routing;
