+++ src/Servers/Servers.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const Servers = () => {
+  return (<div>Servers list</div>)
+}
